feat: notification widget

+++ packages/ui/lib/src/theme/be_style_value.dart
@@ -44,5 +44,5 @@ abstract class BeStyleValue {
 
   // Notification
   double get notificationMaxWidth;
-  double get notificationMaxCount;
+  int get notificationMaxCount;
 }
